move karaokeIcon.cs file into sprite namespace. follow how `OsuIcon` do.
